refactor: rename components and update import paths for better organization in Settings
