+++ frontend/src/components/Settings/Administration/Tasks.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import Task from "@/components/Settings/TaskOption.vue";
+import Task from "@/components/Settings/Administration/TaskOption.vue";
 import RSection from "@/components/common/RSection.vue";
 import api from "@/services/api/index";
 import storeHeartbeat from "@/stores/heartbeat";
