+++ frontend/src/components/Settings/Administration/Users/Dialog/DeleteUser.vue
