+++ frontend/src/components/Settings/LibraryManagement/AddBtn.vue
