+++ frontend/src/components/Settings/LibraryManagement/Dialog/CreatePlatformBinding.vue
