+++ frontend/src/components/Settings/LibraryManagement/Dialog/CreatePlatformVersion.vue
