+++ frontend/src/components/Settings/LibraryManagement/Dialog/DeletePlatformBinding.vue
