+++ frontend/src/components/Settings/LibraryManagement/Dialog/DeletePlatformVersion.vue
