+++ frontend/src/components/Settings/LibraryManagement/ExcludedCard.vue
