+++ frontend/src/components/Settings/LibraryManagement/PlatformBindCard.vue
