+++ frontend/src/views/Settings/Administration.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
-import UsersTable from "@/components/Administration/Users/Table.vue";
-import Tasks from "@/components/Administration/Tasks.vue";
+import UsersTable from "@/components/Settings/Administration/Users/Table.vue";
+import Tasks from "@/components/Settings/Administration/Tasks.vue";
 import storeAuth from "@/stores/auth";
 
 // Props
