+++ frontend/src/views/Settings/LibraryManagement.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
-import Excluded from "@/components/Management/Excluded.vue";
-import PlatformBinding from "@/components/Management/PlatformBinding.vue";
-import PlatformVersions from "@/components/Management/PlatformVersions.vue";
+import Excluded from "@/components/Settings/LibraryManagement/Excluded.vue";
+import PlatformBinding from "@/components/Settings/LibraryManagement/PlatformBinding.vue";
+import PlatformVersions from "@/components/Settings/LibraryManagement/PlatformVersions.vue";
 </script>
 
 <template>
